requirements.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5

# Vector Database
chromadb==0.4.22
faiss-cpu==1.7.4

# Embedding Models
sentence-transformers==2.2.2

# PDF Processing
pdfplumber==0.10.0
PyPDF2==3.0.1

# OpenAI Integration
openai==1.7.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Web Interface
streamlit==1.29.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
requests>=2.31.0

# Persistence
pickle-mixin==1.0.2

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Logging
loguru==0.7.2

# Optional: For advanced embeddings
transformers>=4.30.0
torch>=2.0.0  # Only if you want to use local transformers

# Web Interface (Phase 5)
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0

scipy>=1.10.0
faiss-cpu>=1.7.0  # Alternative to ChromaDB

# Optional: For better text processing
nltk>=3.8.0
spacy>=3.7.0